sel: add a function to atomically get and clear a SEL entry
This function allows the user to extract SEL entries without the risk
of the entry changing in between getting and clearing it.

Before this change, we were able to call `get_sel_entries()` and
`clear_sel()` after, but there was a risk of missing the entries that
were added during the time between those two calls.

Signed-off-by: Emilio Perez <emilio.perez-juarez@diamond.ac.uk>